[modernize] Use strong types for all quotes calculations, use strong types in pricing

+++ web-gui/flaskr/model/quote.py
@@ -1,12 +1,13 @@
 from pydantic import BaseModel, Field, HttpUrl
 from typing import List, Optional
+from decimal import Decimal
 from datetime import datetime
 from .types import PyObjectId
 
 
 class QuoteHistoryItem(BaseModel):
     timestamp: datetime
-    quote: float
+    quote: Decimal
 
 
 class QuoteCurrencyPair(BaseModel):
